Refactor members global access to make sure all members obj are same
rancher/rancher#18278
