+++ lib/global-admin/addon/components/new-multi-cluster-app/component.js
@@ -209,28 +209,11 @@ export default Component.extend(NewOrEdit, CatalogApp, {
 
     addAuthorizedPrincipal(principal) {
       if (principal) {
-        let { members } = this.multiClusterApp;
-        const { principalType, id } = principal;
+        let { members = [] } = this.multiClusterApp;
 
-        const nue = {
-          type:        'member',
-          accessType:  null,
-          displayType: get(principal, 'displayType') || principalType,
-          displayName: get(principal, 'displayName') || get(principal, 'loginName') || get(principalType, 'id'),
-        };
+        members.pushObject(principal);
 
-        if (!members) {
-          members = [];
-        }
-
-        if (principalType === 'group') {
-          set(nue, 'groupPrincipalId', id);
-        } else if (principalType === 'user') {
-          set(nue, 'userPrincipalId', id);
-        }
-
-
-        members.pushObject(nue);
+        set(this, 'multiClusterApp.members', members);
       }
     },
