+++ lib/global-admin/addon/global-dns/entries/new/controller.js
@@ -91,21 +91,10 @@ export default Controller.extend(ViewNewEdit, {
     addAuthorizedPrincipal(principal) {
       if (principal) {
         let { members = [] } = this.model.globaldns;
-        const { principalType, id } = principal;
 
-        const nue = {
-          type:        'member',
-          accessType:  'owner',
-        };
+        set(principal, 'accessType', 'owner');
 
-        if (principalType === 'group') {
-          set(nue, 'groupPrincipalId', id);
-        } else if (principalType === 'user') {
-          set(nue, 'userPrincipalId', id);
-        }
-
-
-        members.pushObject(this.globalStore.createRecord(nue));
+        members.pushObject(this.globalStore.createRecord(principal));
 
         set(this, 'model.globaldns.members', members);
       }
