+++ lib/global-admin/addon/global-dns/providers/new/controller.js
@@ -30,23 +30,10 @@ export default Controller.extend(NewOrEdit, {
     addAuthorizedPrincipal(principal) {
       if (principal) {
         let { members = [] } = this.model;
-        const { principalType, id } = principal;
 
-        const nue = {
-          type:        'member',
-          accessType:  'owner',
-          displayType: get(principal, 'displayType') || principalType,
-          displayName: get(principal, 'displayName') || get(principal, 'loginName') || get(principalType, 'id'),
-        };
+        set(principal, 'accessType', 'owner');
 
-        if (principalType === 'group') {
-          set(nue, 'groupPrincipalId', id);
-        } else if (principalType === 'user') {
-          set(nue, 'userPrincipalId', id);
-        }
-
-
-        members.pushObject(nue);
+        members.pushObject(this.globalStore.createRecord(principal));
 
         set(this, 'model.members', members);
       }
