+++ lib/shared/addon/components/form-members-global-access/component.js
@@ -1,6 +1,6 @@
 import Component from '@ember/component';
 import layout from './template';
-import { set } from '@ember/object';
+import { get, set } from '@ember/object';
 import { inject as service } from '@ember/service';
 
 const MEMBERS_HEADERS = [
@@ -31,7 +31,6 @@ export default Component.extend({
   descending:             false,
   ownerOnly:              false,
   resource:               null,
-  addAuthorizedPrincipal: null,
   gotError:               null,
   removeMember:           null,
   optionsForAccessType:   null,
@@ -42,6 +41,33 @@ export default Component.extend({
     if (!this.ownerOnly) {
       this.initOptionsForMembersAccessType();
     }
+  },
+
+  actions: {
+    addPrincipal(principal) {
+      if (principal) {
+        const { principalType, id } = principal;
+
+        const nue = {
+          type:        'member',
+          accessType:  null,
+          displayType: get(principal, 'displayType') || principalType,
+          displayName: get(principal, 'displayName') || get(principal, 'loginName') || get(principalType, 'id'),
+        };
+
+        if (principalType === 'group') {
+          set(nue, 'groupPrincipalId', id);
+        } else if (principalType === 'user') {
+          set(nue, 'userPrincipalId', id);
+        }
+
+        this.addAuthorizedPrincipal(nue);
+      }
+    },
+  },
+
+  addAuthorizedPrincipal() {
+    throw new Error('add principal handler must be provided!!');
   },
 
   initOptionsForMembersAccessType() {
