+++ lib/shared/addon/components/form-members-global-access/template.hbs
@@ -4,7 +4,7 @@
     {{input-identity
       allowTeams=true
       allowLocal=true
-      action=addAuthorizedPrincipal
+      action="addPrincipal"
       onError=gotError
     }}
   </div>
